+++ .gitignore
@@ -65,3 +65,4 @@ local.properties
 .settings/
 target/*
 *.properties
+/target/
