--- src/main/java/controller/HomeController.java
@@ -1,13 +1,0 @@
-package controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class HomeController {
-	
-	@RequestMapping("/dessertfarm.com")
-	public String home() {
-		return "home/homePage";
-	}
-}
