+++ src/main/java/kr/co/dessertfarm/config/MvcConfig.java
@@ -1,4 +1,4 @@
-package config;
+package kr.co.dessertfarm.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.DefaultServletHandlerConfigurer;
