+++ src/main/java/kr/co/dessertfarm/manager/JoinManagerRequest.java
@@ -1,15 +1,16 @@
-package login;
+package kr.co.dessertfarm.manager;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class ManagerRequest {
+public class JoinManagerRequest {
 	private int manager_num;
 	private String manager_name;
 	private String manager_id;
 	private String manager_pwd;
+	private String confirm_manager_pwd;
 	private String manager_email;
 	private String manager_address;
 	private String manager_tel;
