+++ src/main/java/kr/co/dessertfarm/manager/ManagerRequest.java
@@ -1,16 +1,15 @@
-package join;
+package kr.co.dessertfarm.manager;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class JoinManagerRequest {
+public class ManagerRequest {
 	private int manager_num;
 	private String manager_name;
 	private String manager_id;
 	private String manager_pwd;
-	private String confirm_manager_pwd;
 	private String manager_email;
 	private String manager_address;
 	private String manager_tel;
