+++ src/main/webapp/WEB-INF/views/admin/adminJoin.jsp
@@ -2,7 +2,6 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
 <%@ taglib prefix="form" uri="http://www.springframework.org/tags/form" %>
-<%@ page import="spring.*" %>   
 <!DOCTYPE html>
 <html>
 <head>
