+++ src/main/webapp/WEB-INF/views/admin/adminLogin.jsp
@@ -1,4 +1,3 @@
-<%@page import="spring.MainDAO"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
